Update quan ly san va lich dat san

+++ QLSanBong/DTO/LoaiSan.cs
@@ -11,12 +11,12 @@ namespace QLSanBong.DTO
     {
         public int MaLoai { get; set; }
         public string TenLoai { get; set; }
-        public double GiaThue { get; set; }
+        public decimal GiaThue { get; set; }
         public LoaiSan(DataRow row)
         {
             this.MaLoai = (int)row["MaLoai"];
             this.TenLoai = row["TenLoai"].ToString();
-            this.GiaThue = (double)row["GiaThue"];
+            this.GiaThue = (decimal)row["GiaThue"];
         }
 
         public LoaiSan() { }
